requirements.txt
python-telegram-bot>=21.0
python-dotenv>=1.0.0
Flask>=2.0.0

# Dependências de teste (não usadas em produção)
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
